Move files to public folder, and implements Turbo frames
When you click the link and the response from the server arrives, Turbo finds the <turbo-frame> element with the matching ID, and replaces the current <turbo-frame> element with the one from the response.

The <turbo-frame id="gallery"> element on the index.html page is replaced with the matching turbo frame element that arrives from the response. That's how the image + link is replaced with a new image + link.